Moved search_box into objects/.
