+++ lib/map/search_box.js
@@ -1,5 +1,5 @@
 const Settings = require('../settings');
-const SearchBox = require('../search_box');
+const SearchBox = require('../objects/search_box');
 
 module.exports = map => {
   map.searchBox = new SearchBox({
