+++ lib/nests/search_box.js
@@ -1,4 +1,4 @@
-const SearchBox = require('../search_box');
+const SearchBox = require('../objects/search_box');
 
 module.exports = map => {
   map.searchBox = new SearchBox({
