+++ lib/objects/search_box.js
